+++ projector_installer/dialogs.py
@@ -13,9 +13,9 @@ import click
 
 from .apps import get_installed_apps, get_app_path, get_compatible_app_names
 from .run_config import get_run_configs, RunConfig, get_run_config_names, \
-    get_used_projector_ports, get_used_http_ports, is_password_protected
+    get_used_projector_ports, is_password_protected
 
-from .global_config import DEF_HTTP_PORT, DEF_PROJECTOR_PORT
+from .global_config import DEF_PROJECTOR_PORT
 from .secure_config import generate_token
 from .utils import get_local_addresses
 
@@ -245,12 +245,6 @@ def get_def_port(ports: List[int], default: int) -> int:
     return port
 
 
-def get_def_http_port() -> int:
-    """Returns unused port for http server."""
-    http_ports = get_used_http_ports()
-    return get_def_port(http_ports, DEF_HTTP_PORT)
-
-
 def get_def_projector_port() -> int:
     """Returns unused port for projector server."""
     ports = get_used_projector_ports()
@@ -267,33 +261,6 @@ def check_listening_address(address: str) -> bool:
     return address in get_all_addresses()
 
 
-def select_http_address(default: str) -> str:
-    """Selects address for http listening."""
-    while True:
-        res: str = click.prompt('Enter HTTP listening address (press ENTER for default)',
-                                default=default)
-
-        if res == '*':
-            res = '0.0.0.0'
-
-        if check_listening_address(res):
-            return res
-
-        click.echo('You entered incorrect address, please try again.')
-        click.echo('You can try to use one of these addresses:')
-
-        for addr in ['*'] + get_all_addresses():
-            click.echo(addr)
-
-
-def select_http_port() -> int:
-    """Selects port for http server."""
-    port = get_def_http_port()
-    res: int = click.prompt('Enter a desired HTTP port (press ENTER for default)',
-                            default=str(port))
-    return res
-
-
 def select_projector_port() -> int:
     """Selects port for projector server."""
     port = get_def_projector_port()
@@ -305,7 +272,7 @@ def select_projector_port() -> int:
 def select_password(prompt: str, default: str = '') -> str:
     """Prompts for password if needed """
     password: str = click.prompt(text=prompt, default=default, hide_input=True,
-                                           confirmation_prompt=True)
+                                 confirmation_prompt=True)
 
     return password
 
@@ -336,12 +303,6 @@ def edit_config(config: RunConfig) -> RunConfig:
     prompt = 'Enter the path to IDE (press ENTER for default)'
     config.path_to_app = click.prompt(prompt, default=config.path_to_app)
 
-    prompt = 'Enter a HTTP listening address (press ENTER for default)'
-    config.http_address = click.prompt(prompt, default=config.http_address)
-
-    prompt = 'Enter a HTTP port (press ENTER for default)'
-    config.http_port = click.prompt(prompt, default=str(config.http_port))
-
     prompt = 'Enter a Projector port (press ENTER for default)'
     config.projector_port = click.prompt(prompt, default=str(config.projector_port))
 
@@ -370,8 +331,6 @@ def make_run_config(config_name: str, app_path: Optional[str] = None) -> RunConf
         print('IDE was not selected, exiting...')
         sys.exit(1)
 
-    http_address = select_http_address('localhost')
-    http_port = select_http_port()
     projector_port = select_projector_port()
     secure_config = click.prompt(
         'Use secure connection '
@@ -382,19 +341,16 @@ def make_run_config(config_name: str, app_path: Optional[str] = None) -> RunConf
     password, ro_password = select_password_pair()
 
     return RunConfig(config_name, expanduser(app_path), projector_port,
-                     http_address, http_port, token, password, ro_password)
+                     token, password, ro_password)
 
 
 class UserInstallInput:
     """Represents user answers during install session"""
 
     # pylint: disable=too-many-instance-attributes
-    def __init__(self, config_name: str, http_address: str, http_port: int,
-                 projector_port: int, do_run: bool, secure_config: bool,
+    def __init__(self, config_name: str, projector_port: int, do_run: bool, secure_config: bool,
                  password: str, ro_password: str) -> None:
         self.config_name: str = config_name
-        self.http_address: str = http_address
-        self.http_port: int = http_port
         self.projector_port: int = projector_port
         self.do_run = do_run
         self.secure_config = secure_config
@@ -409,8 +365,6 @@ def get_user_install_input(config_name_hint: str, auto_run: bool) -> Optional[Us
     if not config_name:
         return None
 
-    http_address = select_http_address('localhost')
-    http_port = select_http_port()
     projector_port = select_projector_port()
     do_run = True if auto_run else click.prompt('Would you like to run installed ide? [y/n]',
                                                 type=bool)
@@ -422,7 +376,7 @@ def get_user_install_input(config_name_hint: str, auto_run: bool) -> Optional[Us
 
     password, ro_password = select_password_pair()
 
-    return UserInstallInput(config_name, http_address, http_port, projector_port,
+    return UserInstallInput(config_name, projector_port,
                             do_run, secure_config, password, ro_password)
 
 
@@ -430,4 +384,4 @@ def make_config_from_input(inp: UserInstallInput) -> RunConfig:
     """Makes run config from user input"""
     token = generate_token() if inp.secure_config else ''
     return RunConfig(inp.config_name, '', inp.projector_port,
-                     inp.http_address, inp.http_port, token, inp.password, inp.ro_password)
+                     token, inp.password, inp.ro_password)